Add support for reading the MDNS (#1)
* Adds support for reading the MDNS

For the integration into Home Assistant, entities need to have a unique
ID. IP address host or device name are not accepted. The `mdns` field
returned by `/esp/get_identification` seams to be the MAC address
prefixed with "ihs_". MAC addresses are accepted as uniqe IDs.

* Changes reading of identification data

to be robust if MDNS field is missing.
Additionally the method is renamed to match it's
behaviour and the log output is more precise.

---------

Co-authored-by: Kai <1528540+lordyavin@users.noreply.github.com>